comment、reply

+++ src/main/java/com/sht/eurasiaring/entity/Comment.java
@@ -16,12 +16,12 @@ public class Comment implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer commentId;
+	private Long commentId;
 	private String content;
 	private String createTime;
 	private Integer number;
-	private Integer userId;  //留言的发布人id
-	private Integer matterId;
+	private Long userId;  //留言的发布人id
+	private Long matterId;
 	private Integer leaf;  //null 用来区分留言和回复
 
 	@Transient
